sanitized codes

+++ react-tasks/src/components/Task.js
@@ -1,13 +1,16 @@
 // src/components/Task.js
-import React, { useMemo, useState, useEffect } from 'react';
+
+import React, { useMemo, useState } from 'react';
+import { Link } from 'react-router-dom';
+
 import { MaterialReactTable } from 'material-react-table';
 import { Box, Button, Typography, IconButton } from '@mui/material';
-import ModalStructure from "./ModalStructure";
-
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
 import UpdateIcon from '@mui/icons-material/Update';
-import { Link } from 'react-router-dom';
+
+import ModalStructure from "./ModalStructure";
+import WeatherTable from './WeatherCaller';
 
 const Task = props => {
   const columns = useMemo(
@@ -39,21 +42,19 @@ const Task = props => {
     ], [],
   );
 
-  const [tasks, setTasks] = props.functions;
-
-  // Start Modal Functionality Block
-  const [show, setShow] = useState(false);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
-
   const BASE_MODAL = {
     title: "Success",
     body: "Error - Should not see this body.",
     btnName: "Close",
     route: "/viewtasks"
   }
+
+  const [tasks, setTasks] = props.functions;
+  const [show, setShow] = useState(false);
   const [modalInfo, setModalInfo] = useState(BASE_MODAL);
-  // End Modal Functionality Block
+
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
 
   function updatePriority() {
     fetch(`http://localhost:8080/api/priority/list`, {
@@ -73,21 +74,17 @@ const Task = props => {
           });
           handleShow();
         } else if (res.status === 400) {
-          return res.json()
-        } else if (res.status === 404) {
+          return res.json();
         } else {
           return Promise.reject(
             new Error(`Unexpected status code: ${res.status}`)
           );
-        }
+        };
       })
-
       .catch(console.error);
   }
 
-
   function deleteTask(taskID) {
-
     fetch(`http://localhost:8080/api/${taskID}`, {
       method: 'DELETE'
     })
@@ -101,9 +98,7 @@ const Task = props => {
           });
           handleShow();
         } else if (res.status === 400) {
-          res.json();
-        } else if (res.status === 404) {
-          console.log('Task ID not found.');
+          return res.json();
         } else {
           return Promise.reject(
             new Error(`Unexpected status code: ${res.status}`)
@@ -113,20 +108,9 @@ const Task = props => {
       .catch(console.error);
   }
 
-  // Need to modify ModalStructure for a second button that only appears sometimes
-  // function deleteConfirm(taskInfo) {
-  //   setModalInfo({
-  //     title: "Delete Confirmation.",
-  //     body: `Are you sure you want to delete ${taskInfo.taskName} from your task list?`,
-  //     btnName: "Close",
-  //     route: "/viewtasks"
-  //   });
-  //   handleShow();
-  // }
-
   return (
     <>
-      {< ModalStructure show={show} handleClose={() => handleClose()} modalInfo={modalInfo} />}
+      < ModalStructure show={show} handleClose={() => handleClose()} modalInfo={modalInfo} />
 
       <MaterialReactTable
         autoResetPageIndex={false}
@@ -141,13 +125,8 @@ const Task = props => {
 
         renderRowActions={({ row, table }) => (
           <Box sx={{ display: 'flex', flexWrap: 'nowrap', gap: '8px' }}>
-
-            <IconButton
-              color="secondary"
-              onClick={() => {
-                table.setEditingRow(row);
-              }}
-            >
+            
+            <IconButton color="secondary" onClick={() => { table.setEditingRow(row) }}>
               <Link to={`/updatetask/${row.original.taskID}`} >
                 <EditIcon />
               </Link>
@@ -156,6 +135,11 @@ const Task = props => {
             <IconButton
               color="error"
               onClick={() => {
+                // TODO Do I really want this crappy confirm window?
+                if (!window.confirm(`Are you sure you want to delete ${row.original.taskName}?`)) {
+                  return;
+                }
+
                 deleteTask(row.original.taskID)
                 tasks.splice(row.index, 1);
 
@@ -167,33 +151,29 @@ const Task = props => {
             >
               <DeleteIcon />
             </IconButton>
-
           </Box>
         )}
 
+        //TODO fix formatting here from grid to something else
         renderDetailPanel={({ row }) => (
-          <Box
-            sx={{
-              display: 'grid',
-              margin: 'auto',
-              gridTemplateColumns: '1fr 1fr',
-              width: '100%',
-            }}
-          >
-            <Typography>Extended Weather API PlaceHolder</Typography>
+          <Box sx={{ display: 'grid', margin: 'auto', gridTemplateColumns: '1fr 1fr', width: '100%' }}>
+
+            <Typography>Weather Information</Typography>
+            <Typography></Typography>
+            {/* {row.original.gplaceID === null ? "No Commute Info" : <CommuteCaller originLat={36.1716} originLong={115.1391} destLat={row.original.gPlaceLat} destLong={row.original.gPlaceLong} />} // TODO clean this up */}
+            {row.original.gPlaceLat == 0.0 ? "No Weather Info" : <WeatherTable lat={row.original.gPlaceLat} long={row.original.gPlaceLong} />}
+            {/* {row.original.gplaceID === null ? "No Location Info" : <LocationInfo gplaceID={row.original.gplaceID} />}  // TODO clean this up */}
+
           </Box>
         )}
 
         renderTopToolbarCustomActions={({ table }) => (
           <Box sx={{ display: 'flex', gap: '1rem', p: '0.5rem', flexWrap: 'wrap' }}>
-            <Button
-              color="primary"
-              onClick={updatePriority}
-              startIcon={<UpdateIcon />}
-              variant="contained"
-            >
+
+            <Button color="primary" onClick={updatePriority} startIcon={<UpdateIcon />} variant="contained">
               Save List Rankings
             </Button>
+
           </Box>
         )}
